animehub_backend/requirements.txt
Flask
Flask-JWT-Extended
mongoengine
python-dotenv